@@ -1,10 +1,11 @@
 {
+	"type": "module",
 	"scripts": {
 		"build": "tsc && webpack"
 	},
 	"devDependencies": {
-		"typescript": "^4.2.3",
-		"webpack": "^5.30.0",
+		"typescript": "^4.2.4",
+		"webpack": "^5.33.2",
 		"webpack-cli": "^4.6.0"
 	}
 }
